# gen/swag-doc/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "gin-http服务文档",
        "title": "gin-http API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:3013",
    "basePath": "/goodsCenterLogic",
    "paths": {
        "/asynq/v1/addAggTask": {
            "get": {
                "description": "asynq-添加聚合任务",
                "tags": [
                    "asynq"
                ],
                "summary": "asynq-添加聚合任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境变量,默认线上; sandbox 沙箱环境, production 生产环境",
                        "name": "env",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.GroupDeliveryTaskAddResponse"
                        }
                    }
                }
            }
        },
        "/asynq/v1/addTask": {
            "get": {
                "description": "asynq-异步任务,可通过： http://localhost:7013/monitoring/ 查看dashbord报表",
                "tags": [
                    "asynq"
                ],
                "summary": "asynq-add异步任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境变量,默认线上; sandbox 沙箱环境, production 生产环境",
                        "name": "env",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.AsynqEmailDeliveryTaskAddResponse"
                        }
                    }
                }
            }
        },
        "/auth/v1/token/generate": {
            "get": {
                "description": "jwt-token生成及校验",
                "tags": [
                    "auth"
                ],
                "summary": "jwt-token生成及校验",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境变量,默认线上; sandbox 沙箱环境, production 生产环境",
                        "name": "env",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "UserID 用户id",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserName 用户名  example:张三",
                        "name": "userName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.AppJwtTokenSwgResponse"
                        }
                    }
                }
            }
        },
        "/v1/common/generateId": {
            "get": {
                "description": "生成id-描述",
                "tags": [
                    "公共接口"
                ],
                "summary": "雪花ID生成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证信息 eg:xxxx-xxxx-xxxx-xxx",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "生成id数量 1-1000",
                        "name": "num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_servicev1.HttpGenerateIDResponse"
                        }
                    }
                }
            }
        },
        "/v1/exampleGet": {
            "get": {
                "description": "get接口示例",
                "tags": [
                    "Example"
                ],
                "summary": "get接口示例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "create_time",
                        "name": "create_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query_id",
                        "name": "query_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_id 用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.ExampleGetResponse"
                        }
                    }
                }
            }
        },
        "/v1/exampleGetOne": {
            "get": {
                "description": "getOne接口示例",
                "tags": [
                    "Example"
                ],
                "summary": "getOne接口示例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "create_time",
                        "name": "create_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query_id",
                        "name": "query_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user_id 用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.ExampleGetOneResponse"
                        }
                    }
                }
            }
        },
        "/v1/examplePost": {
            "post": {
                "description": "\nios购买项类型 \u003ca href=\"https://developer.apple.com/documentation/appstoreconnectapi/list_all_in-app_purchases_for_an_app\"\u003e 详情\u003c/a\u003e \u003c/br\u003e \n\n\nandroid订阅 \u003ca href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/monetization.subscriptions/list?hl=zh-cn\"\u003e 详情\u003c/a\u003e \u003c/br\u003e \n\nandroid非订阅\u003ca href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/inappproducts/list?hl=zh-cn\"\u003e 详情\u003c/a\u003e \u003c/br\u003e \n\nandroid订阅产品的类型\u003ca href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/inappproducts?hl=zh-cn#PurchaseType\"\u003e 详情\u003c/a\u003e \u003c/br\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Example"
                ],
                "summary": "post 接口 示例",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.ExamplePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.ExamplePostResponse"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/entity/1"
                            },
                            "Token": {
                                "type": "string",
                                "description": "token"
                            },
                            "Token2": {
                                "type": "string",
                                "description": "token2"
                            }
                        }
                    }
                }
            }
        },
        "/v1/userScore/add": {
            "post": {
                "description": "gorm 添加数据",
                "tags": [
                    "user_score"
                ],
                "summary": "添加数据",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/userScore/del": {
            "post": {
                "description": "gorm 删除数据",
                "tags": [
                    "user_score"
                ],
                "summary": "删除数据",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http.ResponseData"
                        }
                    }
                }
            }
        },
        "/v1/userScore/find": {
            "get": {
                "description": "gorm 查询一条数据",
                "tags": [
                    "user_score"
                ],
                "summary": "查询一条数据",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.UserScoreFindResponse"
                        }
                    }
                }
            }
        },
        "/v1/userScore/list": {
            "get": {
                "description": "gorm 查询列表",
                "tags": [
                    "user_score"
                ],
                "summary": "查询列表",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_dto.UserScoreListResponse"
                        }
                    }
                }
            }
        },
        "/v1/userScore/update": {
            "post": {
                "description": "gorm 更新数据",
                "tags": [
                    "user_score"
                ],
                "summary": "更新数据",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api_dto.AppJwtTokenResponse": {
            "type": "object",
            "properties": {
                "jwtPayload": {
                    "description": "token payload 信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_auth_application_service.AuthPayload"
                        }
                    ]
                },
                "token": {
                    "description": "Token jwt token",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.AppJwtTokenSwgResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "$ref": "#/definitions/internal_api_dto.AppJwtTokenResponse"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.AsynqEmailDeliveryTaskAddResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "description": "数据data"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.ExampleGetOneResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "$ref": "#/definitions/internal_api_dto.UserPortraitData"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.ExampleGetResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "$ref": "#/definitions/internal_api_dto.UserPortraitData"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.ExamplePostRequest": {
            "type": "object",
            "required": [
                "Title",
                "app_id",
                "question"
            ],
            "properties": {
                "Title": {
                    "description": "标题",
                    "type": "string"
                },
                "app_id": {
                    "description": "应用id",
                    "type": "string"
                },
                "question": {
                    "description": "问题",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.ExamplePostResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "$ref": "#/definitions/internal_api_dto.UserPortraitData"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.GroupDeliveryTaskAddResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "description": "数据data"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.UserPortraitData": {
            "type": "object",
            "properties": {
                "country": {
                    "description": "国家",
                    "type": "string"
                },
                "last_login": {
                    "description": "上次登陆时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "string"
                },
                "vip_info": {
                    "description": "是否为VIP，0/1",
                    "type": "integer"
                }
            }
        },
        "internal_api_dto.UserScoreFindResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "$ref": "#/definitions/internal_module_mockv2_application_entity.UserScore"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_dto.UserScoreListResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_module_mockv2_application_entity.UserScore"
                    }
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_api_http.ResponseData": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "description": "数据data"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        },
        "internal_module_auth_application_service.AuthPayload": {
            "type": "object",
            "properties": {
                "aud": {
                    "type": "string"
                },
                "audience": {
                    "description": "Audience 当前时间",
                    "type": "integer"
                },
                "exp": {
                    "type": "integer"
                },
                "expiresAt": {
                    "description": "ExpiresAt  token 过期时间",
                    "type": "integer"
                },
                "iat": {
                    "type": "integer"
                },
                "iss": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "nbf": {
                    "type": "integer"
                },
                "sub": {
                    "type": "string"
                },
                "uid": {
                    "description": "UID 用户id",
                    "type": "string"
                },
                "username": {
                    "description": "Username 用户名",
                    "type": "string"
                }
            }
        },
        "internal_module_mockv2_application_entity.UserScore": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "score_result": {
                    "type": "integer"
                },
                "update_time": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "zero_timestamp": {
                    "type": "integer"
                }
            }
        },
        "internal_api_http_servicev1.HttpGenerateIDResponse": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "traceId"
            ],
            "properties": {
                "code": {
                    "description": "code:  0 成功; 非0失败;",
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "description": "生成id数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "traceId",
                    "type": "string"
                }
            }
        }
    }
}